pro,wca: Multi-piece mapping submission path support
The provisioning now supports two types of zone management:

(i)  For application-managed metadata, the provisioning returns
     a list of zone offsets and sizes. It forms a multi-piece
     mapping list that is returned to the user application.
     A user buffer can be spread across zones.

(ii) For ZTL-managed metadata we do not support multi-piece mapping
     yet. So, the provisioning returns a single offset and size that
     points to a single zone.
     A user buffer must fit in a single zone.

+++ xapp/src/xapp-mempool.c
@@ -73,7 +73,7 @@ int xapp_mempool_destroy (uint32_t type, uint16_t tid)
     return XAPP_OK;
 }
 
-int xapp_mempool_create (uint32_t type, uint16_t tid, uint16_t entries,
+int xapp_mempool_create (uint32_t type, uint16_t tid, uint32_t entries,
 		uint32_t ent_sz, xapp_mp_alloc *alloc, xapp_mp_free *free)
 {
     struct xapp_mp_pool_i *pool;
